refactor(components): file path changed to ComponentsUsage
